+++ src/types/instagram/feed.ts
@@ -1,0 +1,4 @@
+export interface Feed {
+  id: number;
+  image: string;
+}