+++ src/types/products/attribute.ts
@@ -1,0 +1,6 @@
+export interface Attribute {
+  id: number;
+  name: string;
+  slug: string;
+  type: 'colour' | 'size';
+}