+++ src/types/products/review.ts
@@ -1,0 +1,9 @@
+export interface Review {
+  id: number;
+  productId: number;
+  rating: number;
+  date: string;
+  review: string;
+  name: string;
+  status: string;
+}